requirements.txt
aiofiles==25.1.0
annotated-types==0.7.0
asyncpg==0.31.0
bcrypt==5.0.0
blinker==1.9.0
cffi==2.0.0
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
Flask==3.1.2
greenlet==3.3.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
Hypercorn==0.18.0
hyperframe==6.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
priority==2.0.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
python-dotenv==1.2.1
Quart==0.20.0
SQLAlchemy==2.0.45
sqlmodel==0.0.31
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0
Werkzeug==3.1.5
wsproto==1.3.2